Rename utils module to name_utils for better specificity and avoid naming conflicts.
    Renames utils.py to name_utils.py and test_utils.py to test_name_utils.py to avoid Python import conflicts with existing google_sheet test_utils.py. Updates all imports and patch statements across event.py, players.py, and test files to reference the new module name.
