+++ sources/season_view/google_sheet_view/worksheets/event.py
@@ -7,7 +7,7 @@ from google_sheet import utils as sheet_utils
 from season_common import scorecard
 
 from season_view.api import read_data, write_data
-from season_view.google_sheet_view.worksheets import utils
+from season_view.google_sheet_view.worksheets import name_utils
 
 FTR_WRITER_FORMATTING_ENABLED = True
 
@@ -223,7 +223,7 @@ class EventWorksheetReader:
 
         # Move player names column into the index and process names to normalize them
         worksheet_data.set_index(keys="PLAYER", inplace=True)
-        worksheet_data.index = worksheet_data.index.map(utils.process_raw_player_name)
+        worksheet_data.index = worksheet_data.index.map(name_utils.process_raw_player_name)
 
         return sheet_utils.numericise_all_values(worksheet_data)
